# Speed up named character reference lookup