--- a/source/lexbor/core/sbst.h
+++ b/source/lexbor/core/sbst.h
@@ -17,14 +17,14 @@ extern "C" {
 
 
 typedef struct {
-    lxb_char_t key;
+    lxb_char_t    key;
 
-    void       *value;
-    size_t     value_len;
+    lxb_char_t    value[6];
+    unsigned char value_len;
 
-    size_t     left;
-    size_t     right;
-    size_t     next;
+    unsigned int  left;
+    unsigned int  right;
+    unsigned int  next;
 }
 lexbor_sbst_entry_static_t;