Looking up named character references in the BST takes a long time. This patch specializes the BST data structure a bit to allow faster lookups.

There are two big changes:
1) Don't use void*, but use char[6]. On 64-bit this uses less space, and
   saves a memory dereference.
2) Use smaller data types than `size_t`. We use `unsigned int` instead
   which allows for 4 billion entries, which should be enough for a
   static binary search tree.

Using the benchmark from https://gist.github.com/nielsdos/221e88539e516710f8ac44f6fdde7a79 I obtain the following results on an i7-4790:

* Original:
Benchmark 1: ./parse_old
  Time (mean ± σ):     378.9 ms ±   7.6 ms    [User: 375.8 ms, System: 2.6 ms]
  Range (min … max):   373.9 ms … 399.8 ms    10 runs

* After optimization:
Benchmark 1: ./parse_new
  Time (mean ± σ):     335.7 ms ±   5.7 ms    [User: 334.0 ms, System: 1.3 ms]
  Range (min … max):   330.6 ms … 350.0 ms    10 runs



Now, I understand if you don't want this because it specializes the data structure for named character lookups. Still I think at least the change from `size_t` to `unsigned int` makes sense and should reduce the binary size and may reduce cache pressure for closeby entries.