@lexborisov 2024-08-07
@nielsdos 

Interesting patch.
I love any patch that improves performance. 
Apparently, the structure you suggest fits better in the processor cache.
This is a global change, I need time to study it and analyze it.